+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added support for Builtin keywords of Robot Framework like Run Keyword If, Run Keywords, etc. Highlighting the parameters as keywords or python expressions
   accordingly. Additionally, providing code completion for those parameters
 - Added new inspection to convert deprecated keyword calls into native Robot syntax (i.e. `Run Keyword If` into if statement)
+- Added new inspection to detect duplicated parameter definitions in the same keyword call
 
 ### Changed